Switch rust test install base to /data/local/tmp for consistency with cc
Previously this code was not doing anything because the template used
for rust did not use the template string that gets replaced with the
folder coming from this code. So all tests were getting deployed into
/data/local/tmp regardless of being vendor tests.

The other change in the chain adds the template string, so this code
starts working and system and vendor tests start being deployed into
different folders.

We want to keep the system test folder as before to avoid potential
problems with tests (see b/350479879#comment9).

Bug: 347370677
Test: presubmit
Change-Id: I93d5a8db3470d1603e26f9cf5d12249f29141a19