Append vintf_fragment_modules as required for AndroidMk
Current vintf_fragment_modules are not exposed to Kati, so modules from
the property is not being installed to the image while Kati collects
files to be installed from their dependencies. This change marks modules
from vintf_fragment_modules as required, so their dependencies can be
exposed to the Kati.

Bug: 322089980
Test: aosp cf build succeeded
Change-Id: I66fc8f9ee8c0b797a0c200b7070b83ec9bb0f616